# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KnowFlow AI",
    "description": "Enterprise Knowledge & SOP Agent backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chat/query": {
      "post": {
        "tags": [
          "Chat & RAG"
        ],
        "summary": "Query Knowledge Base",
        "description": "Executes a grounded question-answering query against authorized company documentation.\nEnforces server-side authorization: workspace_id, user_id, and access permissions are derived\nexclusively from the verified user_context.",
        "operationId": "query_knowledge_base_api_chat_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/chat/conversations": {
      "get": {
        "tags": [
          "Chat & RAG"
        ],
        "summary": "List Conversations",
        "description": "Lists conversations belonging to the authenticated user in the current workspace.",
        "operationId": "list_conversations_api_chat_conversations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationRead"
                  },
                  "type": "array",
                  "title": "Response List Conversations Api Chat Conversations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/chat/conversations/{conversation_id}": {
      "get": {
        "tags": [
          "Chat & RAG"
        ],
        "summary": "Get Conversation",
        "description": "Retrieves full conversation with messages and verified citations.",
        "operationId": "get_conversation_api_chat_conversations__conversation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetailRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Chat & RAG"
        ],
        "summary": "Delete Conversation",
        "description": "Deletes a conversation and its messages.",
        "operationId": "delete_conversation_api_chat_conversations__conversation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/feedback": {
      "post": {
        "tags": [
          "Chat & RAG"
        ],
        "summary": "Submit Feedback",
        "description": "Submits positive or negative feedback on an AI answer.",
        "operationId": "submit_feedback_api_chat_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackReadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/chat/query": {
      "post": {
        "tags": [
          "Chat & RAG"
        ],
        "summary": "Query Knowledge Base",
        "description": "Executes a grounded question-answering query against authorized company documentation.\nEnforces server-side authorization: workspace_id, user_id, and access permissions are derived\nexclusively from the verified user_context.",
        "operationId": "query_knowledge_base_api_v1_chat_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chat/conversations": {
      "get": {
        "tags": [
          "Chat & RAG"
        ],
        "summary": "List Conversations",
        "description": "Lists conversations belonging to the authenticated user in the current workspace.",
        "operationId": "list_conversations_api_v1_chat_conversations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationRead"
                  },
                  "type": "array",
                  "title": "Response List Conversations Api V1 Chat Conversations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chat/conversations/{conversation_id}": {
      "get": {
        "tags": [
          "Chat & RAG"
        ],
        "summary": "Get Conversation",
        "description": "Retrieves full conversation with messages and verified citations.",
        "operationId": "get_conversation_api_v1_chat_conversations__conversation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetailRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Chat & RAG"
        ],
        "summary": "Delete Conversation",
        "description": "Deletes a conversation and its messages.",
        "operationId": "delete_conversation_api_v1_chat_conversations__conversation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/feedback": {
      "post": {
        "tags": [
          "Chat & RAG"
        ],
        "summary": "Submit Feedback",
        "description": "Submits positive or negative feedback on an AI answer.",
        "operationId": "submit_feedback_api_v1_chat_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackReadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatQueryRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "description": "The user's query text."
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id",
            "description": "Existing conversation ID to append message to."
          },
          "department_filter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Filter",
            "description": "Optional department filter to narrow query within user's authorized scope."
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatQueryRequest"
      },
      "ChatQueryResponse": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "user_message_id": {
            "type": "string",
            "title": "User Message Id"
          },
          "assistant_message_id": {
            "type": "string",
            "title": "Assistant Message Id"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "is_grounded": {
            "type": "boolean",
            "title": "Is Grounded"
          },
          "policy_applied": {
            "type": "string",
            "title": "Policy Applied",
            "default": "Grounded Answering Policy"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationRead"
            },
            "type": "array",
            "title": "Citations"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms",
            "default": 0.0
          },
          "tokens_used": {
            "type": "integer",
            "title": "Tokens Used",
            "default": 0
          },
          "model_used": {
            "type": "string",
            "title": "Model Used"
          },
          "grounding_status": {
            "type": "string",
            "title": "Grounding Status"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "user_message_id",
          "assistant_message_id",
          "query",
          "answer",
          "is_grounded",
          "model_used",
          "grounding_status"
        ],
        "title": "ChatQueryResponse"
      },
      "CitationRead": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title"
          },
          "page_number": {
            "type": "integer",
            "title": "Page Number",
            "default": 1
          },
          "section_heading": {
            "type": "string",
            "title": "Section Heading",
            "default": ""
          },
          "relevant_text": {
            "type": "string",
            "title": "Relevant Text",
            "default": ""
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "document_title"
        ],
        "title": "CitationRead"
      },
      "ConversationDetailRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageRead"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "user_id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ConversationDetailRead"
      },
      "ConversationRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "user_id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ConversationRead"
      },
...
      "FeedbackCreateRequest": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": -1.0,
            "title": "Rating",
            "description": "Thumbs down (-1), thumbs up (1), or star rating (1-5)"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "rating"
        ],
        "title": "FeedbackCreateRequest"
      },
      "FeedbackReadResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "RECORDED"
          }
        },
        "type": "object",
        "required": [
          "id",
          "message_id",
          "rating"
        ],
        "title": "FeedbackReadResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "sender_type": {
            "type": "string",
            "title": "Sender Type"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/CitationRead"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "id",
          "conversation_id",
          "sender_type",
          "content",
          "created_at"
        ],
        "title": "MessageRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}